generic: pid parameter added
Now the Netlink class and all its children have the `pid`
parameter in constructors.

1. pid == None: let the library to choose pid + port value
2. pid == 0: use os.getpid() only, no port allocation
3. pid == <int>: set the pid to some custom value

The first approach allows to use several netlink sockets for
the same family within one process. It is good mainly for RTNL
solutions.

The second is useful for generic netlink, and it is required
by NETLINK_FIREWALL.