rtnl: support IFLA_NET_NS_FD nla
To specify netns, one can use two ways:
1. net_ns_fd = int
2. net_ns_fd = 'name'

In the first case, the class doesn't care about fd and just provides the
value to the kernel as is. It is up to the programmer, to open and to
close the fd.

In the second case, the class opens netns from '/var/run/netns' and
closes on the destruction.

FIXME: check fd leaks and implement forced close.

Bug-Url: https://github.com/svinota/pyroute2/issues/84