global: use mixin classes — prepare for netns
NetNS should provide all the same functionality, as IPRoute does. But the main
difference of NetNS will be the lowest socket level: all `recv()`, `sendto()` etc.
methods will be proxied to a separate process, running in some network
namespace. So, this patch provides all the inheritance chain up to IPRoute
implemented with mixin classes — to make it possible to reuse the code for
NetNS.

Bug-Url: https://github.com/svinota/pyroute2/issues/84