netns: alpha-version
Not documented and not well tested yet, though already able to be used. To
run NetNS::

	from pyroute2.netns import NetNS  # later will be changed
	netns = NetNS('test')  # test — netns name
	netns.get_links()
	netns.close()

Since NetNS starts a separate process, the `close()` call is mandatory to
clean up all the library stack.

Please notice, that IPDB can be run with NetNS, but the `release()` still has
bugs and will be fixed later::

	from pyroute2 import IPDB
	from pyroute2.netns import NetNS

	netns = NetNS('test')
	ip = IPDB(nl=netns)
	print(ip.interfaces.keys())
	ip.release()  # has bugs — doesn't work now

Please be aware, that you should have root access to run the library in a
netns. If you haven't, library will silently conrinue to work in a main netns.
This behaviour will be fixed later.

Bug-Url: https://github.com/svinota/pyroute2/issues/84