common: provide a way to define own SocketBase class
The `socket.socket` class is not sublcass-friendly, and sometimes it is
better to use a custom wrapper providing socket API, than the original
socket class.

But some projects, that use pyroute2, already have their own solutions,
and providing the library-wide wrapper breaks the behaviour of these
projects.

So we provide a way to define a custom `SocketBase` class, that will be
used as base for the `NetlinkSocket`.

Bug-Url: https://github.com/svinota/pyroute2/pull/94
Bug-Url: https://github.com/eventlet/eventlet/issues/198