common: provide basic docs in the root proxy
There are two ways to import modules from pyroute2.

* `from pyroute2 import <symbol>`
or
* `from pyroute2.full.path import <symbol>`

The first way allows do not rely on the module layout: the layout can change,
but "root" imports will stay the same. `from pyroute2 import IPRoute` will
always be `from pyroute2 import IPRoute`.

Also, "root" imports do not provide the required class immediately. Instead
they provide a proxy, that loads module only on demand. That's not about
the startup speed, but it allows to safely replace some base classes via
`pyroute2.config`, and only after that load the module.

It is not a monkey-patching, though looks like that.

This patch provides not only proxies, but also a basic docstring to the proxies.
Please note, that referring the docstring by itself also causes the module
load.