nlsocket: register cleanup callbacks by nlmsg
Network namespace assignment requires to open a file before sending
the request. The issue is how to close the open file, since `__del__()`
doesn't help here for some reason.

So this approach is to provide a possibility to explicitly register cleanup
callbacks by nl messages upon encoding, and run them once on the
response.

Bug-Url: https://github.com/svinota/pyroute2/issues/89