ipdb: ignore default dummy0 race
If the required interface is created externally during commit(), don't
fail and just continue to use it.

The typical use case is when one tries to create a `dummy0` interface
when the `dummy` module is not loaded. The module creates `dummy0`
by default, thus acting as an external race condition source.

Bug-Url: https://github.com/svinota/pyroute2/issues/123