AddrPool: `setaddr()` method
`setaddr()` can set an address allocated or free. The difference
from ordinary `alloc()` and `free()` is that it does it
unconditionally and silently ignores if the address is in the
target state already.

The method can be used to ban/unban specific addresses within the
range or to customize the allocation policy.