ipdb: do not use IPRoute monitor mode
The monitor mode leads either to double message parsing — a performance
issue — or it confuses GC, so nlmsg objects stay referred and cause a
memory leak. The most convenient way would be to deprecate the monitor
mode at all.

This change is not yet stable, though it passes all the functional tests.

Bug-Url: https://github.com/svinota/pyroute2/issues/161