Mask SIGINT from netns child processes
By default, pressing Ctrl-C on the keyboard when running a python
process in the foreground sends SIGINT to all processes in the
hierarchy, which results in the cleanup+wait of the parent waiting on
some message from the child, which is also cleaning itself up. To avoid
this, mask SIGINT in the children. This places a burden on the parent to
always wait on the child, but avoids the deadlock.

Signed-off-by: Brenden Blanco <bblanco@plumgrid.com>