Makefile: `uninstall` target
Works correctly only when installed with pip or with
`make install setuplib=setuptools`

The setuptools module is not provided by the stdlib,
it is not the default choice in the Makefile.

The pure distutils module, which is the default, doesn't
leave enough information to track which files should be
removed by the uninstall command.