Add tc clsact and bpf filter direct-action support
In linux 4.5, a new clsact qdisc type is added, which is very similar to
ingress but can also apply to egress packets in a lightweight (lockless)
way.

Also improving performance is the direct-action flag for bpf filters, in
which a program directly returns the action (drop, pass, redirect, etc.)
to take in the classifer.