Fix typo.
This may be the cause for getting no MAC address in the snippet
bellow:

  import pyroute2
  ip = pyroute2.IPDB()
  neigh = ip.interfaces['eth0]['neighbours']
  neigh['192.168.1.10'].get('lladdr')

Assuming a neighbor with IP = 192.168.1.10 connected to the 'eth0'
interface, this would return the neighbor's MAC address.
But it returns 'None'.

The typo origin seems to be in revision dbb0128d, from 2016-05-01,
still at 'ipdb/main.py'.
It then was moved to 'ipdb/interface.py' in revision 64f9cfec,
from 2016-06-02.