eventlet: use bytes literal newline in _MpConnection.readline.
b'\n' is not equal to '\n' in Python 3. The loop in readline() never
ends as sock.recv returns bytes in Python 3 but the readline loop stops
only if string literal '\n' is met. Therefore readline exhausts input
buffer beyond b'\n' and that hangs the whole program on sock.recv().