ipset: support more ipset types
We already can create ipsets of several types, but that is not enough
to really use them. Indeed, we must support all entries types in the add and
delete commands.

This patch begins the support of other types than IPv4/IPv6 (like iface
and net) and add the support of ipsets with several dimensions
(hash:net,iface for example).

It adds some complexity to parse entries before to add it, since we must
use the good attribute value. Since we are not aware of the ipset type
when we try to insert an entry, the function caller must pass the type
as string (default is "ip", it does not break previous implementation).