ipset: use last revision available by default
Each ipset module as a revision list. This revision number is very
important since:
 * it changes the feature list supported by the kernel for this type
 * kernel returns one error if someone set a bad revision

For example, on a kernel 3.16, last revision for hash:netiface type is
6, when last revision of hash:ip is 4. We don't really can use a good
number for both.

With this patch, we will read the last revision available on the kernel
and use it by default. For compatibility, we keep the attr_revision
variable in __init__, even if it does not really make sense in my
opinion (since this value depends on ipset type, that should probably be
a parameter of create() method)