ipset: rework of structures to match real ipset protocol
We had several issues in this file. The first one is that
IPSET_ATTR_DATA can have several structures (it depends on the command
ran by users) (see lib/PROTOCOL of ipset sources for more information).

The second issue is the structure of IPSET_ATTR_ADT. This nla actually
contains only IPSET_ATTR_DATA structures today. The previous code was
working because IPSET_ATTR_PROTO was used as "recursive". That was bad.

So, this commit introduces a deep rework of ADT and DATA structures:

 * introduce a get_data_type helper to match cadt/adt structures
 * remove unused values in cadt structure, and add "create specific"
   attributes
 * extract ipset_generic structure to share it on several levels

This commit can sadly breaks previous applications based on previous
code, since IPSET_ATTR_ADT as not the same structure anymore. On a
"list" command, the caller was writing this kind of code before:

  ndmsg.get_attr("IPSET_ATTR_ADT").get_attrs("IPSET_ATTR_PROTO")

now, it is:

  ndmsg.get_attr("IPSET_ATTR_ADT").get_attrs("IPSET_ATTR_DATA")

The relevant line in our tests is updated