ipdb: do not drop failed transactions
WARNING: this commit seriously changes the `commit()` behavior
and may cause backwards compatibility issues in some corner
cases.

Previously, the transaction's life cycle was not well defined.
The transaction object was always destroyed after a successful
commit, but in the case of errors there were differences --
sometimes the transaction was destroyed (dropped), sometimes
not.

This commit is the first in the series of changes to introduce
a simple uniform logic.

1. transaction is dropped after a successful commit ...
2. ... unless it was specified manually.
3. transaction always stays after an error and rollback

This change touches only the interfaces plugin.

Saving the transaction after an error gives a chance to fix
the changes and try to apply them again.

At the same time, the issue that is introduced with the commit,
is that any change after a failed commit may go into the same
transaction that caused the error. Thus the code that employed
the implicit transaction drop after an error may face an error.