nlsocket: don't store a reference to the underlying socket methods (they can change)
Use a dynamic __getattr__ based approach instead, to always use the
current socket method.

The reference-based method caused IPRoute sockets to be in an
inconsistent state when closed:
- the socket methods are replaced by dummies at closing time
- the stored reference still points to the old method
- the underlying os-level socket is not necessarily closed
Thus data still could be sent over the socket even if it was considered
closed.

Only Python 2 is affected, because in Python 3 socket methods point to
the "real thing" and are not replaced by dummies when closing.