Fix IPRoute.get() lockup due to internal exceptions
The method defined in netlink.nlsocket.NetlinkMixin.get() does not
cleanly release locks when an exception is encountered internally (for
example, an exception when calling recv() on the socket, etc). The
result is that an IPRoute instance can enter a locked up state in which
it is unable to receive any more messages via the get() method.