rtnl: fix wrong rtmgrp constant names (#471)
The RTNLGRP_ prefix is reserved for multicast group ids (to be used
with setsockopt()). The group flags (legacy api to be used with bind())
are named with the RTMGRP_ prefix.

Here is a link to the original linux commit that introduced the new api:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=ac6d439d2097

Rename the constants to use the correct prefix. And rename RTNL_GROUPS
to RTMGRP_DEFAULTS for consistency.

Fix the map_namespace prefix to RTM_ to avoid accidentally merging
RTMGRP_ constants in RTM_NAMES and RTM_VALUES.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>