IPDB: introduce eventqueue()/nextmsg() interface
Event queue interface is an alternative to "post" callbacks.
While "post" callbacks are executed in a separate thread each,
event queue interface follows the "eventloop" metaphor, i.e.
each netlink event received in the monitoring thread is put in
the queue from which it can be subsequently fetched by calling
`nextmsg()` generator function in the main thread (or in any other
thread(s) started by the user). In the event of packet storm, it
is much nicer to the resources than creating a thread for each
received message.

Signed-off-by: Eugene Crosser <crosser@average.org>