Use threading.local to Store Cleanup Callback Dict
The clean_cbs dict in netlink/__init__.py uses seq as the key. When
pyroute2 is used in multithread programe, each thread get's its own
AddrPool and the seq always starts from 0x000000ff. If clean_cbs is not
stored in thread local storage, it's possible for threads happen to get
the same seq and register cleanup under the same key. So if one thread
does its cleanup, it frees the resources owned by other threads.

In this patch we store the cleanup table in thread local storage to
avoid the problem. Other global variables in netlink/__init__.py don't
suffer similar race conditions, so no need to change.

Bug-Url: https://github.com/svinota/pyroute2/issues/547
Signed-off-by: Zhengsheng Zhou <zhengshengz@vmware.com>