Stop NetNS server gracefully
This patch changes the way of stopping the NetNS server, by sending
a SIGTERM signal and waiting for the server loop to finish. When the
signal is received, the loop control flag is inverted. Once the loop
is finished, the server process ends.

This patch also modifies the Transport class receiver function loops.
When the Transport object is closed, the receiver loops are stopped but
not the file descriptors. Those file descriptors, created in the NetNS
parent class, are closed at the end of the NetNS.close function, once
the child process (Server) is finished and the Transport receiver loops
are stopped. At this point, the file descriptors are not in use and can
be closed.

closes #578