ndb: SQLite3 request cache workaround
The NDB module may corrupt the SQLite3 request cache so
it completely stops working, returning KeyError.

When the SQL request cache is full, SQLite3 module removes
one request from it in order to cache the new one. The
KeyError exception contains then the SQL request that
must be removed from the cache, but the removal failed.

Key functions (may be used to set up breakpoints in gdb):
* Modules/_sqlite/cursor.c:pysqlite_query_execute
* Modules/_sqlite/cache.c:pysqlite_cache_get