ndb: change the DB locking scheme
The problem: pysqlite3 does not work with multiple threads. Even
if the locking is implemented on the higher level, there are
still issues with the SQL queries cache, that may be hidden, but
nobody knows when it blows up.

Here we try to sacrifice the performance in favor of the
reliability -- the DB will be accessed from one thread only,
all the communication will go transparently through the
management thread.

See also: 5b13e563d6daec0a511addad1dd382c490b9f1da