wiset: do not raise exception on recent kernels
On "old" kernels, like version 4.19, the kernel does not return any
message for an invalid ipset name:

In [2]: load_ipset("doesnotexists")
In [3]:

However, on recent kernel, we now receive an error. It's probably a good
thing, but it breaks our code compatibility here. So we know catch the
exception to keep consistent behaviour between kernel versions