noipdb: transitional module
A transitional module that aims to help to move the code from
IPDB to NDB.

NoIPDB API is designed to be a drop-in replacement for IPDB,
but instead of just replacing the obsolete module, it prints
the differences.

Another goal of NoIPDB is to apply IPDB practices to NDB and
implement the missing parts.

The initial commit does not include neighbours and routes
management, only basic interfaces management is provided yet.