iproute: add strict_check option
strict_check option enables some features in kernel. Obviously, it
returns more errors/sanitization of netlink messages. But it is in same
time mandatory for feature like kernel side filtering of IP addresses.

For example, IPRoute already supports to give ifindex in netlink message
for RTM_GETADDR. But it is ignored in kernel, since strict_check is not
set.

This patch can provide a good performance gain in that case:

def perf_get(strict_check):
    with IPRoute(strict_check=strict_check) as ipr:
        len(ipr.addr("dump", index=1, family=AF_INET))

 %timeit perf_get(False)
100 loops, best of 5: 10.9 ms per loop

 %timeit perf_get(True)
1000 loops, best of 5: 210 µs per loop

Tested with ~250 interfaces UP and running one IP address on each.