generic: add support for dumping policy
Add basic support for dumping the policy data that the kernel can report
via CTRL_CMD_GETPOLICY.

The current implementation for parsing the mappings is a bit awkward since
it relies on a list comprehension to fill in the nla_map used.

This is done because the kernel presents some of the sections with nested
NLA where the attribute number is actually an index or a command opcode,
rather than using an extra attribute. Ideally, the nla or nlmsg classes
would be expanded to support this kind of nested substructure.

Eventually we probably want to extract the policy data into more useful
structures. In particular, converting the command opcodes and attribute ids
into human readable names would be ideal.

Add a very basic example which simply dumps and prints the policy data for
the nlctrl netlink family.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>