Add _peek_bufsize to NetlinkSocketBase so recv can use it
NetlinkSocketBase.recv uses DEFAULT_RCVBUF to fetch things like ip links
via netlink sockets. It happed in the last cycles that with new NICs
with higher VF numbers or similar we have to change the buffer size
(see #751 or #813).
A better solution (thanks to one or our experts) can be to check the
necessary buffer size before receiving the sockets.

Fixes #1044

Change-Id: Ide711f27c99e4dfb75fb579f10f005cb8e1b9b37