Allow passing args up to GenericNetlinkSocket constructor
Currently, most subclasses of GenericNetlinkSocket do not pass unused
args up to GenericNetlinkSocket.__init__.

One use case of this is to pass fileno= with an existing file descriptor
that was opened in another process/network namespace.

This patch fixes the following subclasses

  - EventSocket
  - NL80211
  - DevlinkSocket
  - L2tp
  - WireGuard

by passing up *args and **kwargs.

We also fix TaskStats by simply removing the __init__ method that just
calls super.

We did not need to fix MPTCP and NlEthtool as they don't overwrite
__init__.