ethtool: generalize get_stringset to work with different set IDs
Stringset length varies by driver and set_id.

Hack: The ctypes library does not allow us to assign redefined/resized
Gstring structures to the same IfReq because it (rightly) considers a
new structure definition incompatible with the original one. If we were
to comply with strict type checking we would have to redefine IfReq
every time, but that creates new sort of mess.

Instead of redefining IfReq we define gstrings structure member as
void* pointer and that allows us to reassign resized structure without
redefining the whole IfReq every time. The same applies to gstats member.