ci/openstack: use netns by default
+ update Octavia tests

The issue with running tests without a netns is that the main
namespace can have systemd, NetworkManager or something like that,
and creating a new interface in the main namespace triggers them
to do weird things with the interface.

To avoid such interference, it is better not to use the main
network namespace at all.