uevent: Add type annotations
Do not reassign variables to avoid type conflict reports from mypy. Find
a more efficient way for splitting lines into key and value while at it.

Turn parse() on MarshalUevent into a generator so it matches the
superclass' signature.  Not sure if this is still semantically correct,
the return in the error case and Generator SendType and ReturnType in
particular. (my generator-fu is weak)

Add superclass parameters to parse() to match signature (and ignore
them).

Annotating bind() on UeventSocket seems to work because mypy can't
follow the *argv, **kwargs on the superclass back into
AsyncNetlinkSocket.

No such luck on AsyncUeventSocket: While bind()'s signature can be made
compatible with that of AsyncNetlinkSocket, the one of AsyncCoreSocket
is plain incompatible (addr arg). So we basically have to promise mypy
that we know what we're doing at this point. See
https://mypy.readthedocs.io/en/stable/common_issues.html#incompatible-overrides