ethtool/ioctl: Use byref instead of cpointer when possible
According to the ctypes documentation [1], calling cpointer creates
a new instance, and it seems to lead to memory leaks: "Note that ctypes
does not have OOR (original object return), it constructs a new,
equivalent object each time you retrieve an attribute".

The cpointer documentation [2] indicates that byref should rather
be used when passing "a pointer to an object to a foreign function
call".

[1]: https://docs.python.org/3/library/ctypes.html#pointers
[2]: https://docs.python.org/3/library/ctypes.html#ctypes.pointer