netlink: return error message in exception_factory
errmsg contains headers of original message. That is valuable to build
better exception.

Warning: it breaks previous API of exception_factory, adding one
argument. Since this feature does not look really used outside of ipset
module (for now), I think that it is OK. But a breaking change warning
in changelog will be mandatory