ethtool: Add a method to retrieve SFP module information
Use the `MODULE_EEPROM_GET` netlink message to dump the whole module
EEPROM, then parse the EEPROM according to the module type in order
to retrieve module information.

This is basically a porting of the `-m`/`--dump-module-eeprom`/
`--module-info` option processing of `ethtool`:
- The EEPROM dump is an adapation of the `eeprom_parse()` function,
which is located in `netlink/module-eeprom.c` [1];
- The dump parsing is an adaptation of the `sff8079_show_all_nl()`
function, which is located in `sfpid.c` [2].

Please note that only the parsing of SFF-8079 module EEPROM is
implemented, contrary to the parsing of the SFF-8636 and CMIS ones.

[1]: https://git.kernel.org/pub/scm/network/ethtool/ethtool.git/tree/netlink/module-eeprom.c#n210
[2]: https://git.kernel.org/pub/scm/network/ethtool/ethtool.git/tree/sfpid.c#n502