dhcp: Do task cleanup for early exit from detect_servers()
I was trying to use detect_servers() to do a DHCP discovery and exit as
soon as I got the first reply. However, the discovery tasks are never
stopped when you return inside the async for loop on detect_servers().
This commit makes sure that all tasks are always cancelled properly.